+++ packages/client/src/Client.ts
@@ -17,6 +17,7 @@
 import type {
   ActionDefinition,
   InterfaceDefinition,
+  MinimalActionDefinition,
   ObjectTypeDefinition,
   QueryDefinition,
   VersionBound,
@@ -48,7 +49,7 @@ export interface Client extends SharedClient<MinimalClient> {
     o: Q,
   ): Q["objectSet"];
 
-  <Q extends ActionDefinition<any, any, any>>(
+  <Q extends MinimalActionDefinition<any, any, any>>(
     o: Q,
   ): ActionSignatureFromDef<Q>;