+++ packages/internal.foundry.core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @osdk/internal.foundry.core
 
+## 0.2.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [01724ae]
+  - @osdk/shared.net@2.0.0-beta.2
+
 ## 0.2.0-beta.5
 
 ### Patch Changes